test: normalize expected paths for Windows
When testing on Windows, the tool currently prints paths with a `/` path
separator rather than the canonical `\` separator for the platform.  The
resulting path is valid and usable.  However, this causes match failures
due to the difference in the path separator.  The `slash` module permits
us to normalize the path for the tests.  This should have no impact when
running on Unix platforms, but allows greater portability to Windows.