test: support test.utils.command-exists on Windows
`cat` is a common utility on Unix systems, however, it is not
universally available.  On Windows systems, resort to `cmd` (%COMSPEC%)
as the shell interpreter is guaranteed to be present.  It stands to
reason that `sh` would be a viable alternative to `cat` as that is the
default (POSIX) shell interpreter on Unix systems and should be present
on any SuS or POSIX complaint system.