utils: add new `execute` utility
This new function is similar to `run` except that it requires that the
arguments be passed properly quoted and pre-split through an array.  It
also will not permit the prefixing a command for execution.  The major
difference between `run` and `execute` is that `execute` will not go
through a shell, so parameter expansion, variable substitution, etc do
not occur.  Additionally, because there is no shell involved, a single
process is created rather than two processes (one for the shell, and the
actual process being invoked).  This is required to enable using this
action on Windows, where the process redirection does not work as the
shell used is different.